+++ .eslintrc.js
@@ -1,7 +1,7 @@
 // This configuration only applies to the package manager root.
 /** @type {import("eslint").Linter.Config} */
 module.exports = {
-  extends: ['@token-icons/eslint-config/library.js'],
+  extends: ['@web3icons/eslint-config/library.js'],
   parser: '@typescript-eslint/parser',
   parserOptions: {
     project: true,